requirements.txt
streamlit
tensorflow
opencv-python-headless
tf-keras
fpdf2
numpy
Pillow
matplotlib
